+++ src/control_plane/migration/mongodb/mod.rs
@@ -1,0 +1,6 @@
+mod backup_mongodb_database;
+mod mongodb_backup_options;
+mod mongodb_connection_uri;
+
+pub(crate) use backup_mongodb_database::backup_mongodb_database;
+pub(crate) use mongodb_backup_options::MongoDbBackupOptions;